+++ src/which.cpp
@@ -164,8 +164,8 @@ SEXP cpp_which_val(SEXP x, SEXP value, bool invert){
     SEXP out = SHIELD(new_vec(is_long ? REALSXP : INTSXP, out_size));
     ++NP;
     SHIELD(value = coerce_vector(value, CHEAPR_INT64SXP)); ++NP;
-    long long int val = INTEGER64_PTR(value)[0];
-    long long int *p_x = INTEGER64_PTR(x);
+    int_fast64_t val = INTEGER64_PTR(value)[0];
+    int_fast64_t *p_x = INTEGER64_PTR(x);
     if (is_long){
       double *p_out = REAL(out);
       if (invert){
@@ -253,7 +253,7 @@ SEXP cpp_which_na(SEXP x){
   }
   case CHEAPR_INT64SXP: {
     R_xlen_t count = na_count(x, true);
-    long long *p_x = INTEGER64_PTR(x);
+    int_fast64_t *p_x = INTEGER64_PTR(x);
     if (is_short){
       int out_size = count;
       SEXP out = SHIELD(new_vec(INTSXP, out_size));
@@ -415,7 +415,7 @@ SEXP cpp_which_not_na(SEXP x){
   }
   case CHEAPR_INT64SXP: {
     R_xlen_t count = na_count(x, true);
-    long long *p_x = INTEGER64_PTR(x);
+    int_fast64_t *p_x = INTEGER64_PTR(x);
     if (is_short){
       int out_size = n - count;
       SEXP out = SHIELD(new_vec(INTSXP, out_size));
@@ -740,8 +740,8 @@ SEXP cpp_val_find(SEXP x, SEXP value, bool invert, SEXP n_values){
     SEXP out = SHIELD(new_vec(is_long ? REALSXP : INTSXP, out_size));
     ++NP;
     SHIELD(value = coerce_vector(value, CHEAPR_INT64SXP)); ++NP;
-    long long int val = INTEGER64_PTR(value)[0];
-    long long int *p_x = INTEGER64_PTR(x);
+    int_fast64_t val = INTEGER64_PTR(value)[0];
+    int_fast64_t *p_x = INTEGER64_PTR(x);
     if (is_long){
       double *p_out = REAL(out);
       if (invert){
